+++ src/DocumentFragment.php
@@ -6,6 +6,7 @@ namespace Wikimedia\Dodo;
 
 use Wikimedia\Dodo\Internal\NamespacePrefixMap;
 use Wikimedia\Dodo\Internal\UnimplementedTrait;
+use Wikimedia\Dodo\Internal\Util;
 
 /**
  * DocumentFragment
@@ -77,6 +78,49 @@ class DocumentFragment extends ContainerNode implements \Wikimedia\IDLeDOM\Docum
 				$markup
 			);
 		}
+	}
+
+	// Somewhat annoyingly, querySelector/querySelectorAll needs some methods
+	// of Document/Element... but DocumentFragment doesn't have them.
+	// Fake it out by constructing a pseudo-element!
+
+	/** @inheritDoc */
+	public function querySelectorAll( string $selectors ) {
+		$fakeElement = new class( $this ) extends Element {
+			/** @var DocumentFragment $docFrag */
+			private $docFrag;
+
+			/** @param DocumentFragment $docFrag */
+			public function __construct( $docFrag ) {
+				parent::__construct(
+					$docFrag->_nodeDocument,
+					'fake',
+					Util::NAMESPACE_HTML,
+					null
+				);
+				$this->docFrag = $docFrag;
+			}
+
+			// Believe it or not, this is the only method we need to fake
+
+			/** @inheritDoc */
+			public function _nextElement( ?Element $root ): ?Element {
+				return $this->docFrag->getFirstElementChild();
+			}
+		};
+
+		return $fakeElement->querySelectorAll( $selectors );
+	}
+
+	// Just implement querySelector in terms of querySelectorAll
+
+	/** @inheritDoc */
+	public function querySelector( string $selectors ) {
+		$nodeList = $this->querySelectorAll( $selectors );
+		if ( $nodeList->getLength() > 0 ) {
+			return $nodeList->item( 0 );
+		}
+		return null;
 	}
 
 	// Non-standard, but useful (github issue #73)