+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * EXPERIMENTAL: Added Node::getExtensionData() and Node::setExtensionData()
   methods to allow end-users to associate additional non-spec data off of
   Nodes.
+* Implement DocumentFragment::querySelector() and ::querySelectorAll()
 
 # Dodo 0.1.0 (2021-07-04)
 Initial release.